requirements.txt
aerich==0.6.3
beanie==1.11.6
black==22.6.0
fastapi-users-tortoise==0.2.0
fastapi-users[beanie,oauth]==10.1
fastapi-utils==0.2.1
fastapi[all]==0.78.0
inflect==5.6.0
ipython==8.4.0
bpython==0.23
loguru==0.6.0
mypy==0.971
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-cov==3.0.0
pytest-dotenv==0.5.2
python-dotenv==0.20.0
pytz==2022.1
strawberry-graphql[fastapi, debug-server]==0.119.0
tortoise-orm[asyncpg, accel]==0.19.2
uvicorn[standard]<=0.18.0
